validate_token: Move `validate_token` to avoid dependency cycle.
Prep commit to avoid depencency cycle.
